Refactor predictSequence into TensorflowForClassification trait [skip test]
