+++ src/main/scala/com/johnsnowlabs/ml/tensorflow/TensorflowForClassification.scala
@@ -16,7 +16,7 @@
 
 package com.johnsnowlabs.ml.tensorflow
 
-import com.johnsnowlabs.nlp.annotators.common.{IndexedToken, TokenPiece, TokenizedSentence, WordpieceTokenizedSentence}
+import com.johnsnowlabs.nlp.annotators.common.{IndexedToken, Sentence, TokenPiece, TokenizedSentence, WordpieceTokenizedSentence}
 import com.johnsnowlabs.nlp.{Annotation, AnnotatorType}
 
 trait TensorflowForClassification {
@@ -49,6 +49,47 @@ trait TensorflowForClassification {
 
   }
 
+  def predictSequence(tokenizedSentences: Seq[TokenizedSentence], sentences: Seq[Sentence], batchSize: Int, maxSentenceLength: Int,
+                      caseSensitive: Boolean, coalesceSentences: Boolean = false, tags: Map[String, Int]): Seq[Annotation] = {
+
+    val wordPieceTokenizedSentences = tokenizeWithAlignment(tokenizedSentences, maxSentenceLength, caseSensitive)
+
+    /*Run calculation by batches*/
+    wordPieceTokenizedSentences.zip(sentences).zipWithIndex.grouped(batchSize).flatMap { batch =>
+      val tokensBatch = batch.map(x => (x._1._1, x._2))
+      val encoded = encode(tokensBatch, maxSentenceLength)
+      val logits = tagSequence(encoded)
+
+      if (coalesceSentences) {
+        val scores = logits.transpose.map(_.sum / logits.length)
+        val label = tags.find(_._2 == scores.zipWithIndex.maxBy(_._1)._2).map(_._1).getOrElse("NA")
+        val meta = scores.zipWithIndex.flatMap(x => Map(tags.find(_._2 == x._2).map(_._1).toString -> x._1.toString))
+        Array(
+          Annotation(
+            annotatorType = AnnotatorType.CATEGORY,
+            begin = sentences.head.start,
+            end = sentences.head.end,
+            result = label,
+            metadata = Map("sentence" -> sentences.head.index.toString) ++ meta
+          )
+        )
+      } else {
+        sentences.zip(logits).map { case (sentence, scores) =>
+          val label = tags.find(_._2 == scores.zipWithIndex.maxBy(_._1)._2).map(_._1).getOrElse("NA")
+          val meta = scores.zipWithIndex.flatMap(x => Map(tags.find(_._2 == x._2).map(_._1).toString -> x._1.toString))
+          Annotation(
+            annotatorType = AnnotatorType.CATEGORY,
+            begin = sentence.start,
+            end = sentence.end,
+            result = label,
+            metadata = Map("sentence" -> sentence.index.toString) ++ meta
+          )
+        }
+      }
+    }.toSeq
+
+  }
+
   def tokenizeWithAlignment(sentences: Seq[TokenizedSentence], maxSeqLength: Int, caseSensitive: Boolean): Seq[WordpieceTokenizedSentence]
 
   /** Encode the input sequence to indexes IDs adding padding where necessary
@@ -69,6 +110,8 @@ trait TensorflowForClassification {
   }
 
   def tag(batch: Seq[Array[Int]]): Seq[Array[Array[Float]]]
+
+  def tagSequence(batch: Seq[Array[Int]]): Array[Array[Float]]
 
   def calculateSoftmax(scores: Array[Float]): Array[Float] = {
     val exp = scores.map(x => math.exp(x))
